Changed API framework file extensions to TypeScript
The mechanical rename keeps the CommonJS implementation intact while preparing the package for a focused typed ESM conversion.
