+++ packages/api-framework/eslint.config.mjs
@@ -4,8 +4,8 @@ export default nodeLibConfig({
   typescript: false,
   commonjs: true,
   legacyLocalFilenames: true,
-  srcGlobs: ['index.js', 'src/**/*.js'],
-  testGlobs: ['test/**/*.js'],
+  srcGlobs: ['index.js', 'src/**/*.ts'],
+  testGlobs: ['test/**/*.ts'],
   extraTestRules: {
     // Tests intentionally exercise propagation of arbitrary native errors.
     'ghost/ghost-custom/no-native-error': 'off',
