--- packages/api-framework/src/serializers/index.js
@@ -1,13 +1,0 @@
-module.exports = {
-  get handle() {
-    return require('./handle');
-  },
-
-  get input() {
-    return require('./input');
-  },
-
-  get output() {
-    return require('./output');
-  },
-};
