+++ packages/api-framework/src/utils/index.ts
@@ -1,5 +1,5 @@
 module.exports = {
   get options() {
-    return require('./options');
+    return require('./options.ts');
   },
 };
