--- packages/api-framework/src/validators/index.js
@@ -1,9 +1,0 @@
-module.exports = {
-  get handle() {
-    return require('./handle');
-  },
-
-  get input() {
-    return require('./input');
-  },
-};
