+++ packages/api-framework/src/validators/index.ts
@@ -1,0 +1,9 @@
+module.exports = {
+  get handle() {
+    return require('./handle.ts');
+  },
+
+  get input() {
+    return require('./input/index.ts');
+  },
+};
