+++ packages/api-framework/src/validators/input/index.ts
@@ -1,5 +1,5 @@
 module.exports = {
   get all() {
-    return require('./all');
+    return require('./all.ts');
   },
 };
