+++ packages/api-framework/vitest.config.ts
@@ -3,9 +3,9 @@ import { createVitestConfig } from '@internal/cfg-vitest';
 export default createVitestConfig({
   test: {
     globals: true,
-    include: ['test/**/*.test.js'],
+    include: ['test/**/*.test.ts'],
     coverage: {
-      include: ['src/**/*.js'],
+      include: ['src/**/*.ts'],
       thresholds: {
         lines: 90,
         functions: 90,
